Moved from core-java to json module
