+++ json/src/main/java/com/baeldung/jsonjava/CookieDemo.java
@@ -1,4 +1,4 @@
-package com.baeldung.java.json;
+package com.baeldung.jsonjava;
 
 import org.json.Cookie;
 import org.json.JSONObject;
