+++ json/src/main/java/com/baeldung/jsonjava/DemoBean.java
@@ -1,4 +1,4 @@
-package com.baeldung.java.json;
+package com.baeldung.jsonjava;
 
 public class DemoBean {
     private int id;
