+++ json/src/main/java/com/baeldung/jsonjava/HTTPDemo.java
@@ -1,4 +1,4 @@
-package com.baeldung.java.json;
+package com.baeldung.jsonjava;
 
 import org.json.HTTP;
 import org.json.JSONObject;
