+++ json/src/main/java/com/baeldung/jsonjava/JSONObjectDemo.java
@@ -1,4 +1,4 @@
-package com.baeldung.java.json;
+package com.baeldung.jsonjava;
 
 import java.util.HashMap;
 import java.util.Map;
