+++ json/src/test/java/com/baeldung/jsonjava/HTTPIntegrationTest.java
@@ -1,4 +1,4 @@
-package com.baeldung.java.json;
+package com.baeldung.jsonjava;
 
 import static org.junit.Assert.assertEquals;
 import org.json.HTTP;
