+++ json/src/test/java/com/baeldung/jsonjava/JSONTokenerIntegrationTest.java
@@ -1,4 +1,4 @@
-package com.baeldung.java.json;
+package com.baeldung.jsonjava;
 
 import static org.junit.Assert.assertEquals;
